combine trollDelete and adminDelete into delete
